Added elec_current
def elec_current calibrates electrified population in urban and rural (and national).

It checks if there are transofmers or MV lines and uses these first. If there are none (columns are missing) it goes with HV.

It also includes a dummy distance in case calibrated elec does not add up in urban and rural settlements.

Requires five new inputs in specs file (these are added and uplaoded to box)

Many hardcoded instances. Consider rearrangning the code to decrease? Also do we want to calibrate elec in urban and rural or only worry about national as we urban/rural is dependent on our own calibration to befin with